# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agrillion API — rewards & utility ecosystem
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Create a new member account and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Sign in with Membership ID / phone / email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: authRefresh
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Revoke the current refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Current member resolved from the access token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Not authenticated
...
  /smart-units/ledger:
    get:
      operationId: listSmartUnitsLedger
      tags: [smart-units]
      summary: Smart Units ledger entries
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartUnitsLedgerEntry"
  /smart-units/transfer:
    post:
      operationId: transferSmartUnits
      tags: [smart-units]
      summary: Transfer Smart Units to another member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferUnitsBody"
      responses:
        "200":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"
...
  /mart/orders:
    get:
      operationId: listMartOrders
      tags: [mart]
      summary: Member's mart orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MartOrder"
    post:
      operationId: createMartOrder
      tags: [mart]
      summary: Place a mart order (cash, units, or split)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMartOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MartOrder"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string, description: "+234... or 0..." }
        state: { type: string }
        lga: { type: string }
        password: { type: string, minLength: 8 }
        tier: { type: string, enum: [member, premier], default: member }
      required: [fullName, email, phone, state, lga, password]
    LoginRequest:
      type: object
      properties:
        identifier: { type: string, description: "Membership ID, email, or phone" }
        password: { type: string }
      required: [identifier, password]
    RefreshRequest:
      type: object
      properties:
        refreshToken: { type: string }
      required: [refreshToken]
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string, description: "Short-lived JWT (15m)" }
        refreshToken: { type: string, description: "Long-lived rotating token (30d)" }
        accessTokenExpiresAt: { type: string, format: date-time }
        member: { $ref: "#/components/schemas/Member" }
      required: [accessToken, refreshToken, accessTokenExpiresAt, member]
    Member:
      type: object
      properties:
        id: { type: string }
        membershipId: { type: string, description: "e.g. AGP-LA-IKE-000245" }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        state: { type: string }
        lga: { type: string }
        tier: { type: string, enum: [member, premier, partner] }
        joinedAt: { type: string, format: date-time }
        avatarUrl: { type: string, nullable: true }
      required: [id, membershipId, fullName, email, phone, state, lga, tier, joinedAt]
...
    SmartUnitsLedgerEntry:
      type: object
      properties:
        id: { type: string }
        sourceType: { type: string, enum: [utility, transfer_in, transfer_out, mart_redeem, bonus, adjustment] }
        referenceId: { type: string, nullable: true }
        unitsAdded: { type: number }
        unitsUsed: { type: number }
        balanceAfter: { type: number }
        description: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, sourceType, unitsAdded, unitsUsed, balanceAfter, description, createdAt]
    TransferUnitsBody:
      type: object
      properties:
        toMembershipId: { type: string }
        units: { type: number }
        walletPin: { type: string }
        note: { type: string, nullable: true }
      required: [toMembershipId, units, walletPin]
    TransferResult:
      type: object
      properties:
        success: { type: boolean }
        newBalance: { type: number }
        reference: { type: string }
      required: [success, newBalance, reference]
...
    MartOrder:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        productImage: { type: string, nullable: true }
        quantity: { type: integer }
        paymentMethod: { type: string, enum: [cash, units, split] }
        cashPaid: { type: number }
        unitsUsed: { type: number }
        totalNgn: { type: number }
        status: { type: string, enum: [pending, processing, shipped, delivered, cancelled] }
        createdAt: { type: string, format: date-time }
      required: [id, productId, productName, quantity, paymentMethod, cashPaid, unitsUsed, totalNgn, status, createdAt]
    CreateMartOrderRequest:
      type: object
      properties:
        productId: { type: string }
        quantity: { type: integer }
        paymentMethod: { type: string, enum: [cash, units, split] }
        cashPaid: { type: number, nullable: true }
        unitsUsed: { type: number, nullable: true }
      required: [productId, quantity, paymentMethod]
...