# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agrillion API — rewards & utility ecosystem
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from "./member";

export interface AuthResponse {
  /** Short-lived JWT (15m) */
  accessToken: string;
  /** Long-lived rotating token (30d) */
  refreshToken: string;
  accessTokenExpiresAt: Date;
  member: Member;
}