# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agrillion API — rewards & utility ecosystem
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  /** Membership ID, email, or phone */
  identifier: string;
  password: string;
}