# lib/api-zod/src/generated/types/refreshRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agrillion API — rewards & utility ecosystem
 * OpenAPI spec version: 0.1.0
 */

export interface RefreshRequest {
  refreshToken: string;
}