# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agrillion API — rewards & utility ecosystem
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterRequestTier } from "./registerRequestTier";

export interface RegisterRequest {
  fullName: string;
  email: string;
  /** +234... or 0... */
  phone: string;
  state: string;
  lga: string;
  /** @minLength 8 */
  password: string;
  tier?: RegisterRequestTier;
}